moved logger and default commands to separate module
