+++ .gitignore
@@ -5,3 +5,5 @@ __pycache__/loaders.cpython-310.pyc
 data/__pycache__/loaders.cpython-310.pyc
 game/__pycache__/state.cpython-310.pyc
 game/__pycache__/help.cpython-310.pyc
+game/__pycache__/settings.cpython-310.pyc
+utils/__pycache__/logger.cpython-310.pyc
